Compositor: Add fg_FOVCenter uniform
Add an fg_FOVCenter per-view array uniform, containing the central
position relative to the viewport, corresponding to the forward view
direction. This is normally (0.5, 0.5), however for asymmetric
projections, as often used by VR headsets, the projection may well
extend further to the periphery in one direction, i.e. further to the
left for the left eye, and to the right for the right eye.

This will be used with deferred shading to correctly convert a
framebuffer texture coordinate to a view space position when the
projection is asymmetric.