mipmap: Fix assert from mishandling of 1x1 textures
When computeMipmap() computes mipmaps for 1x1 textures,
computeNumberOfMipmapLevels() returns 1, and mipmapOffsets gets a single
entry for the size of the original texture, however that entry is popped
from the back (since there are no more mipmap levels beyond the original
texture) before the front entry is read for the memcpy to copy the
texture into the first level of the mipmap.

Since the mipmapOffsets is a std::vector, in practice the pop_back()
will not result in a reallocation and accessing the front entry will
access the intended memory, however when building with optimisation and
debug (e.g. ArchLinux build flags with no CMAKE_BUILD_TYPE), the front()
call will hit an assert since the vector is empty.

This is fixed by reordering the pop_back() and front() calls so that the
front() will always be valid, which alone is sufficient to avoid the
assert.

Additionally since the generation of mipmaps is entirely redundant for
1x1 textures, an early return is added for this case to avoid even the
generation of a new osg::Image() for the mipmaps.