Compositor: Use node masks to disable passes
Commit cd1e2b08f8b7 ("GLcore: Compositor: Use another technique to
enable/disable rendering of passes") changed the disabling of passes to
clear the osg::Camera's graphics context. Unfortunately this doesn't
work with CullThreadPerCameraDrawThreadPerContext or
DrawThreadPerContext multithreading modes due to OSG not handling the
changing of camera graphics contexts while threading is enabled.

Toggling the camera's node mask appears to function equivalently, works
with all multithreading modes, and has been explicitly recommended as
the most efficient way to do it [1], so lets give that a try.

[1] See osg-users thread "Cannot stop rendering slave camera".