nasal/string.c: fix bogus gcc warning in setlen()
Remove the following warning triggered by the first test in setlen():
“the comparison will always evaluate as ‘true’ for the address of ‘buf’
will never be NULL”. gcc sees the address of a naStr.data.buf used as a
boolean expression operand and thinks this is silly, however the address
in question is never evaluated in this condition due to the two tests on
s->emblen. This commit implements the exact same logic as the original
code, but in a way that doesn't trigger the bogus warning.